Change float to double and remove interaction with configuration file

+++ Game/DTOs/WinnerDto.cs
@@ -6,7 +6,7 @@ namespace Game.DTOs
 	public class WinnerDto
 	{
 		public BasePlayer? Player { get; set; }
-		public float Prize { get; set; }
+		public double Prize { get; set; }
 		public PrizeType PrizeType { get; set; }
 		
 	}
