+++ Game/Players/Concrete/UserPlayer.cs
@@ -4,7 +4,7 @@ namespace Game.Players.Concrete
 {
 	internal class UserPlayer : BasePlayer
 	{
-		public UserPlayer(Guid id, string name, float balance, int number) : base(id, name, balance, number)
+		public UserPlayer(Guid id, string name, double balance, int number) : base(id, name, balance, number)
 		{
 		}
 	}
