+++ Game/Services/Abstract/IPrizeCalculationService.cs
@@ -6,6 +6,6 @@ namespace Game.Services.Abstract
 	public interface IPrizeCalculationService
 	{
 		public List<WinnerDto> CalculatePrize(List<BasePlayer> players);
-		public float CalculateHouseProfit(List<BasePlayer> players, List<WinnerDto> winners);
+		public double CalculateHouseProfit(List<BasePlayer> players, List<WinnerDto> winners);
 	}
 }
